fix(cache): KVC-090 residual evict clone + KVC-085/089/091/092 [codex-unavailable]
KIVI residual buffer (§3.3):
- feat: add FP16 residual buffer for most recent decode tokens
- fix(KVC-090 CRIT): evict_k/v must .clone() before in-place shift
- fix(KVC-092 MED): only route to residual after prefill (k_scale init)
- fix(KVC-091 MED): update get_seq_len() docstring for residual semantics
- fix(KVC-089 MED): forward residual_length param in RoleAwareAsymKVCache

KVC-085: torch_ref decode path scale .to(fp16) dtype fix
generate_loop: pass residual_length from runtime_config

Tests: 6 new test cases (residual noop, buffer fill, eviction,
  clear reset, KVC-090 regression, KVC-092 regression)

Review: D1+D2+D4+D5 sub-agent review passed (Codex unavailable)

+++ src/engine/patch_model.py
@@ -873,12 +873,14 @@ def _fused_forward_impl(
         elif decode_impl == "torch_ref":
             if k_int8_for_ref is None or v_int8_for_ref is None:
                 raise RuntimeError("torch_ref decode for int4 requires materialized INT8 tensors.")
+            # KVC-085 extension: torch_ref path also needs fp32→fp16 scale cast
+            # (same root cause as the Triton fused path fix above).
             attn_output_val = _decode_attn_int8_torch_ref(
                 q_kernel,
                 k_int8_for_ref,
                 v_int8_for_ref,
-                k_scale,
-                v_scale,
+                k_scale.to(torch.float16),
+                v_scale.to(torch.float16),
                 context_lens,
                 sm_scale=sm_scale,
             )
